+++ packages/eyes-api/test/utils/fake-sdk.js
@@ -26,7 +26,12 @@ function makeSDK(settings = {}) {
   }
 
   function isSelector(selector) {
-    return utils.types.isString(selector) || utils.types.has(selector, ['type', 'selector'])
+    return (
+      utils.types.isString(selector) ||
+      utils.types.has(selector, ['type', 'selector']) ||
+      (utils.types.has(selector, ['selector']) && isSelector(selector.selector)) ||
+      utils.types.has(selector, 'fakeSelector')
+    )
   }
 
   function makeManager(config) {