+++ packages/snippets/src/browser/cleanupElementMarkers.js
@@ -1,5 +1,0 @@
-function cleanupElementMarkers([elements]) {
-  elements.forEach(element => element.removeAttribute('data-applitools-marker'))
-}
-
-module.exports = cleanupElementMarkers