+++ packages/snippets/src/browser/getShadowRoot.js
@@ -1,0 +1,4 @@
+function getShadowRoot(element) {
+  return element.shadowRoot
+}
+module.exports = getShadowRoot