+++ packages/types/src/setting.ts
@@ -6,10 +6,11 @@ import {
   AccessibilityLevel,
   Proxy,
 } from './data'
+import {SpecSelector} from './spec-driver'
 
-type RegionReference<TElement, TSelector> = Region | ElementReference<TElement, TSelector>
+type RegionReference<TElement, TSelector> = Region | ElementReference<TElement, TSelector> | SpecSelector<TSelector>
 
-type ElementReference<TElement, TSelector> = TElement | TSelector
+type ElementReference<TElement, TSelector> = TElement | SpecSelector<TSelector>
 
 type FrameReference<TElement, TSelector> = ElementReference<TElement, TSelector> | string | number
 
@@ -59,6 +60,7 @@ export type MatchSettings<TRegion> = {
 export type ScreenshotSettings<TElement, TSelector> = {
   region?: RegionReference<TElement, TSelector>
   frames?: (ContextReference<TElement, TSelector> | FrameReference<TElement, TSelector>)[]
+  shadow?: SpecSelector<TSelector>
   scrollRootElement?: ElementReference<TElement, TSelector>
   fully?: boolean
 }